test(architecture): move logistics runtime http api tests
