+++ test/integration/adapters/controllers/logistics-route-plans-cache-runtime.test.ts
@@ -9,14 +9,14 @@ process.env.SUPABASE_SERVICE_ROLE_KEY ??= "test-service-role-key";
 process.env.DATABASE_URL ??= "postgresql://postgres:postgres@127.0.0.1:5432/postgres";
 process.env.SUPABASE_DB_URL ??= process.env.DATABASE_URL;
 
-const { ENV } = await import("../server/lib/env.ts");
+const { ENV } = await import("../../../../server/lib/env.ts");
 const { logisticsRoutePlansNativeRoutes } = await import(
-  "../server/routes/logistics-route-plans.fastify.ts"
+  "../../../../server/routes/logistics-route-plans.fastify.ts"
 );
 const {
   clearRoutePlanMetricsCache,
   clearRoutePlansCache,
-} = await import("../server/lib/logistics-route-plans-cache.ts");
+} = await import("../../../../server/lib/logistics-route-plans-cache.ts");
 
 const VALID_ORIGIN = "http://localhost:3000";
 const SESSION_TOKEN = "clinic-session-token";
